requirements.txt
aiogram==3.13.0
faster-whisper==1.1.0
sqlalchemy[asyncio]==2.0.36
aiosqlite==0.20.0
python-dotenv==1.0.1
aiofiles==23.2.1
openai==1.54.0
httpx==0.27.2
pydub==0.25.1
